Add commitment helper
